+++ example/lib/gallery/sections/sections.dart
@@ -9,6 +9,7 @@ export 'icon_buttons.dart';
 export 'icons.dart';
 export 'iterable_options.dart';
 export 'key_icons.dart';
+export 'loading_indicators.dart';
 export 'screen_transitions.dart';
 export 'selection_list.dart';
 export 'single_child_scroll_view.dart';