+++ CHANGELOG.md
@@ -10,6 +10,14 @@
  - feat!: add button icon themes
  - refactor: Use NesIcons.someIcon instead of NesIcons.instance.someIcon
  - feat: add `NesInputDialog`
+ - feat: add `NesIcons.hourglassMiddle`
+ - feat: add `NesIcons.hourglassTopFull`
+ - feat: add `NesIcons.hourglassBottomFull`
+ - feat: add `NesHourglassLoadingIndicator`
+ - feat: add `NesTerminalLoadingIndicator`
+ - feat: add `NesJumpingIconsLoadingIndicator`
+ - feat: add `NesMutatingIconsLoadingIndicator`
+ - feat: add `NesPixelRowLoadingIndicator`
 
 # 0.8.0
  - feat: add `NesIcons.leftHand`